move wrap bytes to internal json and remove it from dump
